Add support for mapping configuration files
In case remote maps such as LDAP or MySQL are used, the mapping files
are not tab separated key, value items but instead are common postfix
configuration files where the key and value are separated through the
= character.

Do support these as well by using a different source file that takes care
of the correct formatting.

This way the mapping type can be used more flexible and is not limited to
only support the vmail map type.

Drive-By: Correctly extract file_type when using the postfix proxymap
daemon.